hostapp-update-hooks: Fix boot binaries update hook
Replace the "/tmp" path with "/dev/shm" path for dumping
the content of mtd partitions.

Because the update hooks are running in the update container
and "/tmp" is not bind mounted in the container, a
"No space left on device" error will show up when running the hooks.

Because "/dev" is bind mounted inside the container, we use instead
"/dev/shm" path which is also a tmpfs.

Changelog-entry: hostapp-update-hooks: Fix boot binaries update hook
Signed-off-by: Sebastian Panceac <sebastian@balena.io>